Restructure Hub Auth API Classes
